requirements.txt
langchain
python-dotenv
#ipykernel
pandas 
numpy
langchain_community 
langchain-text-splitters
langchain-openai
langchain_groq
uvicorn
fastapi
langserve